llama3/requirements.txt
torch
fairscale
fire
tiktoken==0.4.0
blobfile
gym
transformers>=4.37.0
paho-mqtt
modelscope>=1.12.0
openai>=1.3.0
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
scikit-learn>=1.1.0
firebase-admin>=6.3.0
